Add files via upload
Homework 2 is about how to use pandas profiling tools for the creation of an HTML file report where we can able to get an overview of the data. For this work, we have used two data sets named red_wine.csv and White_wine.csv for the analysis of the data. For the first task create a pandas profiling report to get the overview of the overall data present in the dataset named red_wine.csv. for the second task used the same data set to find out which model gives the best Accuracy and AUC score for a better comparison of the data present. For the next task, a random forest classifier is used to find the ROC curve. For the next task, the dataset named white_wine.csv has been used to get which model gives the AUC score of the model used.